requirements.txt
Flask==3.0.3
tensorflow==2.21.0
opencv-python==4.10.0.84
google-generativeai>=0.5.0
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
python-dotenv==1.0.1